requirements.txt
blinker==1.9.0
click==8.2.1
dnspython==2.7.0
email_validator==2.2.0
Flask==2.3.3
Flask-Login==0.6.3
Flask-Moment==1.0.6
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
gitdb==4.0.12
GitPython==3.1.37
greenlet==3.2.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
markdown2==2.4.10
MarkupSafe==3.0.2
packaging==25.0
psycopg2-binary==2.9.10
python-dotenv==1.0.0
smmap==5.0.2
SQLAlchemy==2.0.41
typing_extensions==4.14.0
Werkzeug==2.3.7
WTForms==3.0.1
markdown==3.8.2